feat: resource balances feature
