+++ client/apps/eternum-mobile/src/features/resource-balances/index.ts
@@ -1,0 +1,1 @@
+export { useResourceBalances } from "./model/use-resource-balances";
